+++ common/src/main/java/dev/terminalmc/clientsort/mixin/MixinClientPacketListener.java
@@ -32,6 +32,8 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
+import static dev.terminalmc.clientsort.config.Config.options;
+
 @Mixin(ClientPacketListener.class)
 public abstract class MixinClientPacketListener extends ClientCommonPacketListenerImpl {
     protected MixinClientPacketListener(Minecraft client, Connection connection, CommonListenerCookie connectionState) {
@@ -41,6 +43,11 @@ public abstract class MixinClientPacketListener extends ClientCommonPacketListen
     @Inject(method = "handleLogin", at = @At("HEAD"))
     private void onLogin(ClientboundLoginPacket packet, CallbackInfo ci) {
         ClientSort.searchOrderUpdated = false;
+        if (Minecraft.getInstance().getSingleplayerServer() == null) {
+            InteractionManager.setTickRate(options().interactionRateServer);
+        } else {
+            InteractionManager.setTickRate(options().interactionRateClient);
+        }
     }
 
     @Inject(method = "handleSetCarriedItem", at = @At("HEAD"))